remove disrespectful language
I certianly meant no ill-will when I originally wrote this sentence 7 or 8(?) years ago, and was pretty much making fun of myself. However in retrospect it reads as inappropriate to me now, and could be interpreted to make some people feel like their participation in the project is not fully welcomed.

Thus, even though this website hasn't been touched in years, I wanted to address it.

Everyone should feel welcome to participate in lolcommits, regardless of their physical apperance or views on free software licenses.